+++ apps/classroomio-com/src/routes/tools/progress/components/DownloadButton.svelte
@@ -1,0 +1,40 @@
+<script lang="ts">
+  import { toPng } from 'html-to-image';
+  import { nodeStore } from './store';
+  export let isDisabled: boolean;
+  export let isDownloading: boolean;
+  export let text: string = 'Generate Progress Report';
+
+  function convertToPng() {
+    isDownloading = true;
+
+    if ($nodeStore) {
+      toPng($nodeStore)
+        .then((dataUrl) => {
+          const link = document.createElement('a');
+          link.download = 'my-image.png';
+          link.href = dataUrl;
+          link.click();
+        })
+        .catch((error) => {
+          console.error('Oops, something went wrong!', error);
+        })
+        .finally(() => {
+          isDownloading = false;
+        });
+    } else {
+      console.error('Node is not defined');
+    }
+  }
+</script>
+
+<button
+  type="button"
+  disabled={isDownloading || !isDisabled ? true : false}
+  on:click={convertToPng}
+  class=" {isDownloading || !isDisabled ? 'bg-gray-500' : 'bg-[#0233BD]'} {text ===
+  'Generate Progress Report'
+    ? 'block md:hidden'
+    : ''} mt-5 text-white text-xs font-semibold w-full py-3 rounded-md"
+  >{isDownloading ? 'Loading...' : text}</button
+>
